Extract base test case class
